Add information about running tests in multiple PHP versions
Initially I was told via support it wasn't possible to run tests across multiple php versions, but as phpenv is used for managing the active PHP version and there are a lot of different versions installed, this turns out to be super easy. 

The only downside is that these are run in sequence and there's no fancy test matrix that shows the status of the tests across the different versions, but given any failed command marks the build as failed it is perfectly capable of showing you when something is wrong.